requirements.txt
alembic==1.17.1
apispec==6.8.4
attrs==25.4.0
blinker==1.9.0
click==8.3.0
flasgger==0.9.7.1
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
mistune==3.1.4
packaging==25.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
webargs==8.7.1
Werkzeug==3.1.3
